Move createQuery tests into their own folder
