+++ tests/createQuery/NoParam.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import { untrack } from 'svelte';
-	import { createQuery, type LoadResult } from '../src/lib/index.ts';
+	import { createQuery, type LoadResult } from '../../src/lib/index.ts';
 
 	let {
 		states,
@@ -10,7 +10,7 @@
 	}: {
 		states: { value: unknown[] };
 		key: string[];
-		loadingFn: (param: { id: number }) => Promise<LoadResult<unknown, unknown>>;
+		loadingFn: () => Promise<LoadResult<unknown, unknown>>;
 		queryOptions?: {
 			staleTime?: number;
 			initialData?: unknown;
@@ -19,17 +19,13 @@
 
 	const testQuery = createQuery(key, loadingFn, queryOptions);
 
-	let param = $state({ id: 1 });
-
-	const { query, reload } = testQuery(param);
+	const { query, reload } = testQuery();
 
 	$effect(() => {
 		states.value = [...untrack(() => states.value), $state.snapshot(query)];
 	});
 </script>
 
-<button onclick={() => param.id--}>Decrement</button>
-<button onclick={() => param.id++}>Increment</button>
 <button onclick={reload}>Reload</button>
 
 <div>Loading: {query.loading}</div>
